final update for the night
added two new elements water summons a puddle that you can take damage from

updated player weapon and ice spear is kinda borked. will fix later

+++ fractures of reality/Assets/Scripts/gameManager.cs
@@ -145,7 +145,7 @@ public class gameManager : MonoBehaviour
 
     public void UpdateWeaponIconUI()
     {
-        int weapon = (int)playerWeapon.GetCurrentWeapon();
+        int weapon =playerWeapon.GetCurrentWeapon();
         //disable all weapon icons first
         wMagicMissileIcon.SetActive(false);
         wFireballIcon.SetActive(false);
